requirements.txt
fastapi
uvicorn
python-dotenv
supabase
pillow
python-multipart
replicate
google-genai
PyJWT
bcrypt
